Automatize button/links generation
Just fill up the `tweet_data` container with your tweets. One per lne.

The script will load them, split by line, loop through the tweets,
generate a link/anchor element based on the markup from a template,
modify the generated element with the tweet details, and finally
add the button to the list of tweets.

Also display the tweet text as tooltip, when hovering on the buttons.